Plot Phylogenetics tree

Plot Phylogenetics tree

Usage

align_phylo(
  phylo,
  ...,
  ladderize = NULL,
  type = "rectangle",
  center = FALSE,
  tree_type = NULL,
  no_axes = NULL,
  active = NULL,
  size = NULL
)

Arguments

phylo

A phylo object.

...

<dyn-dots> Additional arguments passed to geom_segment().

ladderize

A single string of "left" or "right", indicating whether to ladderize the tree. Ladderizing arranges the tree so that the smallest clade is positioned on the "right" or the "left". By default, NULL means the tree will not be ladderized.

type

A string indicates the plot type, "rectangle" or "triangle".

center

A boolean value. if TRUE, nodes are plotted centered with respect to all leaves/tips in the branch. Otherwise (default), plot them in the middle of the direct child nodes.

tree_type

A single string, one of "phylogram" or "cladogram", indicating the type of tree.

phylogram: Represents a phylogenetic tree where branch lengths indicate evolutionary distance or time.

cladogram: Represents a tree where branch lengths are not used, or the branches do not reflect evolutionary time.

Usually, you don't need to modify this.

no_axes

Logical; if TRUE, removes axes elements for the alignment axis using theme_no_axes(). By default, will use the option- "ggalign.align_no_axes".

active

A active() object that defines the context settings when added to a layout.

size

The relative size of the plot, can be specified as a unit(). Note that for circle_layout(), all size values will be interpreted as relative sizes, as this layout type adjusts based on the available space in the circular arrangement.