Add a plot to annotate a series of ranges of observations

Add a plot to annotate a series of ranges of observations

Usage

align_ranges(
  data = waiver(),
  mapping = aes(),
  ranges = waiver(),
  position = waiver(),
  size = NULL,
  active = NULL
)

Arguments

data

A flexible input that specifies the data to be used

NULL: No data is set.

waiver(): Uses the layout matrix.

A function (including purrr-like lambda syntax) that is applied to the layout matrix, and must return a matrix. If you want to transform the final plot data, please use scheme_data().

A matrix, data frame, or atomic vector.

mapping

Default list of aesthetic mappings to use for plot. If not specified, must be supplied in each layer added to the plot.

ranges

A list of ranges to be annotated. Each range will be represented by a facet panal.

position

Which side the link should be added to? A string containing one or more of "t", "l", "b", and "r". For a horizontal stack_layout(), only l (left) and r (right) can be used. For a vertical stack_layout(), only b (bottom) and t (top) are available.

size

The relative size of the plot, can be specified as a unit.

active

A active() object that defines the context settings when added to a layout.

ggplot2 specification

align_ranges initializes a ggplot.

The data in the underlying ggplot object will contain following columns:

.panel: the panel for the aligned axis. It means x-axis for vertical stack layout (including top and bottom annotation), y-axis for horizontal stack layout (including left and right annotation).

.names (vec_names()) and .index (vec_size()/NROW()): a character names (only applicable when names exists) and an integer of index of the original data.

.row_names and .row_index: the row names and an integer of row index of the original matrix (only applicable if data is a matrix).

.column_names and .column_index: the column names and column index of the original matrix (only applicable if data is a matrix).

value: the actual value (only applicable if data is a matrix or atomic vector).

matrix input will be automatically melted into a long foramted data frame.

Atomic vector will be put in the value column of the data frame.

In the case where the input data is already a data frame, following columns (.panel, .index, .names) are added to the data frame.