Arrange Plots Horizontally or Vertically

The cross_align function aligns observations, and allow different layourt ordering index in a single layout.

Two aliases are provided for convenience:

cross_alignv: A special case of cross_align that sets direction = "vertical".

cross_alignh: A special case of cross_align that sets direction = "horizontal".

Usage

cross_align(data = NULL, direction, ..., theme = NULL, sizes = NA)

cross_alignv(data = NULL, ...)

cross_alignh(data = NULL, ...)

Arguments

data

Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout:

For stack_free, fortify_data_frame() will be used to convert data to a data frame.

For stack_align, fortify_matrix() will be used to convert data to a matrix.

direction

A string indicating the direction of the stack layout, either "horizontal" or "vertical".

...

Additional arguments passed to fortify_data_frame() or fortify_matrix().

theme

A theme() used to render the guides, title, subtitle, caption, margins, patch.title, panel.border, and background. If NULL (default), will inherit from the parent layout.

sizes

A numeric or a unit object of length 3 indicating the relative heights (for direction = "horizontal") or widths (for direction = "vertical"). This is only used if you include a nested quad_layout() in the layout.

See also