Dengrogram x and y coordinates

Dengrogram x and y coordinates

Usage

dendrogram_data(
  tree,
  priority = "right",
  center = FALSE,
  type = "rectangle",
  leaf_pos = NULL,
  leaf_braches = NULL,
  reorder_branches = TRUE,
  branch_gap = NULL,
  root = NULL
)

Arguments

tree

A hclust or a dendrogram object.

priority

A string of "left" or "right". if we draw from right to left, the left will override the right, so we take the "left" as the priority. If we draw from left to right, the right will override the left, so we take the "right" as priority. This is used by align_dendro() to provide support of facet operation in ggplot2.

center

A boolean value. if TRUE, nodes are plotted centered with respect to the leaves in the branch. Otherwise (default), plot them in the middle of all direct child nodes.

type

A string indicates the plot type, "rectangle" or "triangle".

leaf_pos

The x-coordinates of the leaf node. Must be the same length of the number of observations in tree.

leaf_braches

Branches of the leaf node. Must be the same length of the number of observations in tree. Usually come from cutree.

reorder_branches

A single boolean value, indicates whether reorder the provided leaf_braches based on the actual index.

branch_gap

A single numeric value indicates the gap between different branches.

root

A length one string or numeric indicates the root branch.

Value

A list of 2 data.frame. One for node coordinates, another for edge coordinates. node and tree segments edge coordinates contains following columns:

index: the original index in the tree for the current node

label: node label text

x and y: x-axis and y-axis coordinates for current node or the start node of the current edge.

xend and yend: the x-axis and y-axis coordinates of the terminal node for current edge.

branch: which branch current node or edge is. You can use this column to color different groups.

panel: which panel current node is, if we split the plot into panel using facet_grid, this column will show which panel current node or edge is from. Note: some nodes may fall outside panel (between two panels), so there are possible NA values in this column.

.panel: Similar with panel column, but always give the correct branch for usage of the ggplot facet.

panel1 and panel2: The panel1 and panel2 variables have the same functionality as panel, but they are specifically for the edge data and correspond to both nodes of each edge.

leaf: A logical value indicates whether current node is a leaf.

Examples

dendrogram_data(hclust(dist(USArrests), "ave"))
#> $node
#>    index          label         x         y branch  leaf panel .panel
#> 1      9        Florida  1.000000  0.000000   root  TRUE  <NA>   <NA>
#> 2     33 North Carolina  2.000000  0.000000   root  TRUE  <NA>   <NA>
#> 3     NA           <NA>  1.500000 38.527912   root FALSE  <NA>   <NA>
#> 4      5     California  3.000000  0.000000   root  TRUE  <NA>   <NA>
#> 5     20       Maryland  4.000000  0.000000   root  TRUE  <NA>   <NA>
#> 6      3        Arizona  5.000000  0.000000   root  TRUE  <NA>   <NA>
#> 7     31     New Mexico  6.000000  0.000000   root  TRUE  <NA>   <NA>
#> 8     NA           <NA>  5.500000 13.896043   root FALSE  <NA>   <NA>
#> 9     NA           <NA>  4.750000 15.453120   root FALSE  <NA>   <NA>
#> 10    NA           <NA>  3.875000 28.012211   root FALSE  <NA>   <NA>
#> 11     8       Delaware  7.000000  0.000000   root  TRUE  <NA>   <NA>
#> 12     1        Alabama  8.000000  0.000000   root  TRUE  <NA>   <NA>
#> 13    18      Louisiana  9.000000  0.000000   root  TRUE  <NA>   <NA>
#> 14    NA           <NA>  8.500000 15.454449   root FALSE  <NA>   <NA>
#> 15    NA           <NA>  7.750000 16.891499   root FALSE  <NA>   <NA>
#> 16    13       Illinois 10.000000  0.000000   root  TRUE  <NA>   <NA>
#> 17    32       New York 11.000000  0.000000   root  TRUE  <NA>   <NA>
#> 18    NA           <NA> 10.500000  6.236986   root FALSE  <NA>   <NA>
#> 19    22       Michigan 12.000000  0.000000   root  TRUE  <NA>   <NA>
#> 20    28         Nevada 13.000000  0.000000   root  TRUE  <NA>   <NA>
#> 21    NA           <NA> 12.500000 13.297368   root FALSE  <NA>   <NA>
#> 22    NA           <NA> 11.500000 18.417331   root FALSE  <NA>   <NA>
#> 23    NA           <NA>  9.625000 26.363428   root FALSE  <NA>   <NA>
#> 24     2         Alaska 14.000000  0.000000   root  TRUE  <NA>   <NA>
#> 25    24    Mississippi 15.000000  0.000000   root  TRUE  <NA>   <NA>
#> 26    40 South Carolina 16.000000  0.000000   root  TRUE  <NA>   <NA>
#> 27    NA           <NA> 15.500000 21.167192   root FALSE  <NA>   <NA>
#> 28    NA           <NA> 14.750000 28.095803   root FALSE  <NA>   <NA>
#> 29    NA           <NA> 12.187500 39.394633   root FALSE  <NA>   <NA>
#> 30    NA           <NA>  8.031250 44.283922   root FALSE  <NA>   <NA>
#> 31    NA           <NA>  4.765625 77.605024   root FALSE  <NA>   <NA>
#> 32    47     Washington 17.000000  0.000000   root  TRUE  <NA>   <NA>
#> 33    37         Oregon 18.000000  0.000000   root  TRUE  <NA>   <NA>
#> 34    50        Wyoming 19.000000  0.000000   root  TRUE  <NA>   <NA>
#> 35    36       Oklahoma 20.000000  0.000000   root  TRUE  <NA>   <NA>
#> 36    46       Virginia 21.000000  0.000000   root  TRUE  <NA>   <NA>
#> 37    NA           <NA> 20.500000  7.355270   root FALSE  <NA>   <NA>
#> 38    NA           <NA> 19.750000 10.736739   root FALSE  <NA>   <NA>
#> 39    NA           <NA> 18.875000 12.878100   root FALSE  <NA>   <NA>
#> 40    NA           <NA> 17.937500 16.425489   root FALSE  <NA>   <NA>
#> 41    39   Rhode Island 22.000000  0.000000   root  TRUE  <NA>   <NA>
#> 42    21  Massachusetts 23.000000  0.000000   root  TRUE  <NA>   <NA>
#> 43    30     New Jersey 24.000000  0.000000   root  TRUE  <NA>   <NA>
#> 44    NA           <NA> 23.500000 11.456439   root FALSE  <NA>   <NA>
#> 45    NA           <NA> 22.750000 22.595978   root FALSE  <NA>   <NA>
#> 46    NA           <NA> 20.343750 26.713777   root FALSE  <NA>   <NA>
#> 47    25       Missouri 25.000000  0.000000   root  TRUE  <NA>   <NA>
#> 48     4       Arkansas 26.000000  0.000000   root  TRUE  <NA>   <NA>
#> 49    42      Tennessee 27.000000  0.000000   root  TRUE  <NA>   <NA>
#> 50    NA           <NA> 26.500000 12.614278   root FALSE  <NA>   <NA>
#> 51    NA           <NA> 25.750000 20.198479   root FALSE  <NA>   <NA>
#> 52    10        Georgia 28.000000  0.000000   root  TRUE  <NA>   <NA>
#> 53     6       Colorado 29.000000  0.000000   root  TRUE  <NA>   <NA>
#> 54    43          Texas 30.000000  0.000000   root  TRUE  <NA>   <NA>
#> 55    NA           <NA> 29.500000 14.501034   root FALSE  <NA>   <NA>
#> 56    NA           <NA> 28.750000 23.972143   root FALSE  <NA>   <NA>
#> 57    NA           <NA> 27.250000 29.054195   root FALSE  <NA>   <NA>
#> 58    NA           <NA> 23.796875 44.837933   root FALSE  <NA>   <NA>
#> 59    12          Idaho 31.000000  0.000000   root  TRUE  <NA>   <NA>
#> 60    27       Nebraska 32.000000  0.000000   root  TRUE  <NA>   <NA>
#> 61    17       Kentucky 33.000000  0.000000   root  TRUE  <NA>   <NA>
#> 62    26        Montana 34.000000  0.000000   root  TRUE  <NA>   <NA>
#> 63    NA           <NA> 33.500000  3.834058   root FALSE  <NA>   <NA>
#> 64    NA           <NA> 32.750000 12.438692   root FALSE  <NA>   <NA>
#> 65    NA           <NA> 31.875000 15.026107   root FALSE  <NA>   <NA>
#> 66    35           Ohio 35.000000  0.000000   root  TRUE  <NA>   <NA>
#> 67    44           Utah 36.000000  0.000000   root  TRUE  <NA>   <NA>
#> 68    NA           <NA> 35.500000  6.637771   root FALSE  <NA>   <NA>
#> 69    14        Indiana 37.000000  0.000000   root  TRUE  <NA>   <NA>
#> 70    16         Kansas 38.000000  0.000000   root  TRUE  <NA>   <NA>
#> 71    NA           <NA> 37.500000  3.929377   root FALSE  <NA>   <NA>
#> 72     7    Connecticut 39.000000  0.000000   root  TRUE  <NA>   <NA>
#> 73    38   Pennsylvania 40.000000  0.000000   root  TRUE  <NA>   <NA>
#> 74    NA           <NA> 39.500000  8.027453   root FALSE  <NA>   <NA>
#> 75    NA           <NA> 38.500000 13.352260   root FALSE  <NA>   <NA>
#> 76    NA           <NA> 37.000000 15.122897   root FALSE  <NA>   <NA>
#> 77    NA           <NA> 34.437500 20.598507   root FALSE  <NA>   <NA>
#> 78    11         Hawaii 41.000000  0.000000   root  TRUE  <NA>   <NA>
#> 79    48  West Virginia 42.000000  0.000000   root  TRUE  <NA>   <NA>
#> 80    19          Maine 43.000000  0.000000   root  TRUE  <NA>   <NA>
#> 81    41   South Dakota 44.000000  0.000000   root  TRUE  <NA>   <NA>
#> 82    NA           <NA> 43.500000  8.537564   root FALSE  <NA>   <NA>
#> 83    NA           <NA> 42.750000 10.771175   root FALSE  <NA>   <NA>
#> 84    34   North Dakota 45.000000  0.000000   root  TRUE  <NA>   <NA>
#> 85    45        Vermont 46.000000  0.000000   root  TRUE  <NA>   <NA>
#> 86    NA           <NA> 45.500000 13.044922   root FALSE  <NA>   <NA>
#> 87    23      Minnesota 47.000000  0.000000   root  TRUE  <NA>   <NA>
#> 88    49      Wisconsin 48.000000  0.000000   root  TRUE  <NA>   <NA>
#> 89    15           Iowa 49.000000  0.000000   root  TRUE  <NA>   <NA>
#> 90    29  New Hampshire 50.000000  0.000000   root  TRUE  <NA>   <NA>
#> 91    NA           <NA> 49.500000  2.291288   root FALSE  <NA>   <NA>
#> 92    NA           <NA> 48.750000 10.184218   root FALSE  <NA>   <NA>
#> 93    NA           <NA> 47.875000 18.993398   root FALSE  <NA>   <NA>
#> 94    NA           <NA> 46.687500 27.779904   root FALSE  <NA>   <NA>
#> 95    NA           <NA> 44.718750 33.117815   root FALSE  <NA>   <NA>
#> 96    NA           <NA> 42.859375 41.094765   root FALSE  <NA>   <NA>
#> 97    NA           <NA> 38.648438 54.746831   root FALSE  <NA>   <NA>
#> 98    NA           <NA> 31.222656 89.232093   root FALSE  <NA>   <NA>
#> 
#> $edge
#>             x      xend          y       yend branch panel1 panel2 .panel
#> 1    1.000000  1.000000   0.000000  38.527912   root   <NA>   <NA>   <NA>
#> 2    2.000000  2.000000   0.000000  38.527912   root   <NA>   <NA>   <NA>
#> 3    1.500000  1.000000  38.527912  38.527912   root   <NA>   <NA>   <NA>
#> 4    1.500000  2.000000  38.527912  38.527912   root   <NA>   <NA>   <NA>
#> 5    5.000000  5.000000   0.000000  13.896043   root   <NA>   <NA>   <NA>
#> 6    6.000000  6.000000   0.000000  13.896043   root   <NA>   <NA>   <NA>
#> 7    5.500000  5.000000  13.896043  13.896043   root   <NA>   <NA>   <NA>
#> 8    5.500000  6.000000  13.896043  13.896043   root   <NA>   <NA>   <NA>
#> 9    4.000000  4.000000   0.000000  15.453120   root   <NA>   <NA>   <NA>
#> 10   5.500000  5.500000  13.896043  15.453120   root   <NA>   <NA>   <NA>
#> 11   4.750000  4.000000  15.453120  15.453120   root   <NA>   <NA>   <NA>
#> 12   4.750000  5.500000  15.453120  15.453120   root   <NA>   <NA>   <NA>
#> 13   3.000000  3.000000   0.000000  28.012211   root   <NA>   <NA>   <NA>
#> 14   4.750000  4.750000  15.453120  28.012211   root   <NA>   <NA>   <NA>
#> 15   3.875000  3.000000  28.012211  28.012211   root   <NA>   <NA>   <NA>
#> 16   3.875000  4.750000  28.012211  28.012211   root   <NA>   <NA>   <NA>
#> 17   8.000000  8.000000   0.000000  15.454449   root   <NA>   <NA>   <NA>
#> 18   9.000000  9.000000   0.000000  15.454449   root   <NA>   <NA>   <NA>
#> 19   8.500000  8.000000  15.454449  15.454449   root   <NA>   <NA>   <NA>
#> 20   8.500000  9.000000  15.454449  15.454449   root   <NA>   <NA>   <NA>
#> 21   7.000000  7.000000   0.000000  16.891499   root   <NA>   <NA>   <NA>
#> 22   8.500000  8.500000  15.454449  16.891499   root   <NA>   <NA>   <NA>
#> 23   7.750000  7.000000  16.891499  16.891499   root   <NA>   <NA>   <NA>
#> 24   7.750000  8.500000  16.891499  16.891499   root   <NA>   <NA>   <NA>
#> 25  10.000000 10.000000   0.000000   6.236986   root   <NA>   <NA>   <NA>
#> 26  11.000000 11.000000   0.000000   6.236986   root   <NA>   <NA>   <NA>
#> 27  10.500000 10.000000   6.236986   6.236986   root   <NA>   <NA>   <NA>
#> 28  10.500000 11.000000   6.236986   6.236986   root   <NA>   <NA>   <NA>
#> 29  12.000000 12.000000   0.000000  13.297368   root   <NA>   <NA>   <NA>
#> 30  13.000000 13.000000   0.000000  13.297368   root   <NA>   <NA>   <NA>
#> 31  12.500000 12.000000  13.297368  13.297368   root   <NA>   <NA>   <NA>
#> 32  12.500000 13.000000  13.297368  13.297368   root   <NA>   <NA>   <NA>
#> 33  10.500000 10.500000   6.236986  18.417331   root   <NA>   <NA>   <NA>
#> 34  12.500000 12.500000  13.297368  18.417331   root   <NA>   <NA>   <NA>
#> 35  11.500000 10.500000  18.417331  18.417331   root   <NA>   <NA>   <NA>
#> 36  11.500000 12.500000  18.417331  18.417331   root   <NA>   <NA>   <NA>
#> 37   7.750000  7.750000  16.891499  26.363428   root   <NA>   <NA>   <NA>
#> 38  11.500000 11.500000  18.417331  26.363428   root   <NA>   <NA>   <NA>
#> 39   9.625000  7.750000  26.363428  26.363428   root   <NA>   <NA>   <NA>
#> 40   9.625000 11.500000  26.363428  26.363428   root   <NA>   <NA>   <NA>
#> 41  15.000000 15.000000   0.000000  21.167192   root   <NA>   <NA>   <NA>
#> 42  16.000000 16.000000   0.000000  21.167192   root   <NA>   <NA>   <NA>
#> 43  15.500000 15.000000  21.167192  21.167192   root   <NA>   <NA>   <NA>
#> 44  15.500000 16.000000  21.167192  21.167192   root   <NA>   <NA>   <NA>
#> 45  14.000000 14.000000   0.000000  28.095803   root   <NA>   <NA>   <NA>
#> 46  15.500000 15.500000  21.167192  28.095803   root   <NA>   <NA>   <NA>
#> 47  14.750000 14.000000  28.095803  28.095803   root   <NA>   <NA>   <NA>
#> 48  14.750000 15.500000  28.095803  28.095803   root   <NA>   <NA>   <NA>
#> 49   9.625000  9.625000  26.363428  39.394633   root   <NA>   <NA>   <NA>
#> 50  14.750000 14.750000  28.095803  39.394633   root   <NA>   <NA>   <NA>
#> 51  12.187500  9.625000  39.394633  39.394633   root   <NA>   <NA>   <NA>
#> 52  12.187500 14.750000  39.394633  39.394633   root   <NA>   <NA>   <NA>
#> 53   3.875000  3.875000  28.012211  44.283922   root   <NA>   <NA>   <NA>
#> 54  12.187500 12.187500  39.394633  44.283922   root   <NA>   <NA>   <NA>
#> 55   8.031250  3.875000  44.283922  44.283922   root   <NA>   <NA>   <NA>
#> 56   8.031250 12.187500  44.283922  44.283922   root   <NA>   <NA>   <NA>
#> 57   1.500000  1.500000  38.527912  77.605024   root   <NA>   <NA>   <NA>
#> 58   8.031250  8.031250  44.283922  77.605024   root   <NA>   <NA>   <NA>
#> 59   4.765625  1.500000  77.605024  77.605024   root   <NA>   <NA>   <NA>
#> 60   4.765625  8.031250  77.605024  77.605024   root   <NA>   <NA>   <NA>
#> 61  20.000000 20.000000   0.000000   7.355270   root   <NA>   <NA>   <NA>
#> 62  21.000000 21.000000   0.000000   7.355270   root   <NA>   <NA>   <NA>
#> 63  20.500000 20.000000   7.355270   7.355270   root   <NA>   <NA>   <NA>
#> 64  20.500000 21.000000   7.355270   7.355270   root   <NA>   <NA>   <NA>
#> 65  19.000000 19.000000   0.000000  10.736739   root   <NA>   <NA>   <NA>
#> 66  20.500000 20.500000   7.355270  10.736739   root   <NA>   <NA>   <NA>
#> 67  19.750000 19.000000  10.736739  10.736739   root   <NA>   <NA>   <NA>
#> 68  19.750000 20.500000  10.736739  10.736739   root   <NA>   <NA>   <NA>
#> 69  18.000000 18.000000   0.000000  12.878100   root   <NA>   <NA>   <NA>
#> 70  19.750000 19.750000  10.736739  12.878100   root   <NA>   <NA>   <NA>
#> 71  18.875000 18.000000  12.878100  12.878100   root   <NA>   <NA>   <NA>
#> 72  18.875000 19.750000  12.878100  12.878100   root   <NA>   <NA>   <NA>
#> 73  17.000000 17.000000   0.000000  16.425489   root   <NA>   <NA>   <NA>
#> 74  18.875000 18.875000  12.878100  16.425489   root   <NA>   <NA>   <NA>
#> 75  17.937500 17.000000  16.425489  16.425489   root   <NA>   <NA>   <NA>
#> 76  17.937500 18.875000  16.425489  16.425489   root   <NA>   <NA>   <NA>
#> 77  23.000000 23.000000   0.000000  11.456439   root   <NA>   <NA>   <NA>
#> 78  24.000000 24.000000   0.000000  11.456439   root   <NA>   <NA>   <NA>
#> 79  23.500000 23.000000  11.456439  11.456439   root   <NA>   <NA>   <NA>
#> 80  23.500000 24.000000  11.456439  11.456439   root   <NA>   <NA>   <NA>
#> 81  22.000000 22.000000   0.000000  22.595978   root   <NA>   <NA>   <NA>
#> 82  23.500000 23.500000  11.456439  22.595978   root   <NA>   <NA>   <NA>
#> 83  22.750000 22.000000  22.595978  22.595978   root   <NA>   <NA>   <NA>
#> 84  22.750000 23.500000  22.595978  22.595978   root   <NA>   <NA>   <NA>
#> 85  17.937500 17.937500  16.425489  26.713777   root   <NA>   <NA>   <NA>
#> 86  22.750000 22.750000  22.595978  26.713777   root   <NA>   <NA>   <NA>
#> 87  20.343750 17.937500  26.713777  26.713777   root   <NA>   <NA>   <NA>
#> 88  20.343750 22.750000  26.713777  26.713777   root   <NA>   <NA>   <NA>
#> 89  26.000000 26.000000   0.000000  12.614278   root   <NA>   <NA>   <NA>
#> 90  27.000000 27.000000   0.000000  12.614278   root   <NA>   <NA>   <NA>
#> 91  26.500000 26.000000  12.614278  12.614278   root   <NA>   <NA>   <NA>
#> 92  26.500000 27.000000  12.614278  12.614278   root   <NA>   <NA>   <NA>
#> 93  25.000000 25.000000   0.000000  20.198479   root   <NA>   <NA>   <NA>
#> 94  26.500000 26.500000  12.614278  20.198479   root   <NA>   <NA>   <NA>
#> 95  25.750000 25.000000  20.198479  20.198479   root   <NA>   <NA>   <NA>
#> 96  25.750000 26.500000  20.198479  20.198479   root   <NA>   <NA>   <NA>
#> 97  29.000000 29.000000   0.000000  14.501034   root   <NA>   <NA>   <NA>
#> 98  30.000000 30.000000   0.000000  14.501034   root   <NA>   <NA>   <NA>
#> 99  29.500000 29.000000  14.501034  14.501034   root   <NA>   <NA>   <NA>
#> 100 29.500000 30.000000  14.501034  14.501034   root   <NA>   <NA>   <NA>
#> 101 28.000000 28.000000   0.000000  23.972143   root   <NA>   <NA>   <NA>
#> 102 29.500000 29.500000  14.501034  23.972143   root   <NA>   <NA>   <NA>
#> 103 28.750000 28.000000  23.972143  23.972143   root   <NA>   <NA>   <NA>
#> 104 28.750000 29.500000  23.972143  23.972143   root   <NA>   <NA>   <NA>
#> 105 25.750000 25.750000  20.198479  29.054195   root   <NA>   <NA>   <NA>
#> 106 28.750000 28.750000  23.972143  29.054195   root   <NA>   <NA>   <NA>
#> 107 27.250000 25.750000  29.054195  29.054195   root   <NA>   <NA>   <NA>
#> 108 27.250000 28.750000  29.054195  29.054195   root   <NA>   <NA>   <NA>
#> 109 20.343750 20.343750  26.713777  44.837933   root   <NA>   <NA>   <NA>
#> 110 27.250000 27.250000  29.054195  44.837933   root   <NA>   <NA>   <NA>
#> 111 23.796875 20.343750  44.837933  44.837933   root   <NA>   <NA>   <NA>
#> 112 23.796875 27.250000  44.837933  44.837933   root   <NA>   <NA>   <NA>
#> 113 33.000000 33.000000   0.000000   3.834058   root   <NA>   <NA>   <NA>
#> 114 34.000000 34.000000   0.000000   3.834058   root   <NA>   <NA>   <NA>
#> 115 33.500000 33.000000   3.834058   3.834058   root   <NA>   <NA>   <NA>
#> 116 33.500000 34.000000   3.834058   3.834058   root   <NA>   <NA>   <NA>
#> 117 32.000000 32.000000   0.000000  12.438692   root   <NA>   <NA>   <NA>
#> 118 33.500000 33.500000   3.834058  12.438692   root   <NA>   <NA>   <NA>
#> 119 32.750000 32.000000  12.438692  12.438692   root   <NA>   <NA>   <NA>
#> 120 32.750000 33.500000  12.438692  12.438692   root   <NA>   <NA>   <NA>
#> 121 31.000000 31.000000   0.000000  15.026107   root   <NA>   <NA>   <NA>
#> 122 32.750000 32.750000  12.438692  15.026107   root   <NA>   <NA>   <NA>
#> 123 31.875000 31.000000  15.026107  15.026107   root   <NA>   <NA>   <NA>
#> 124 31.875000 32.750000  15.026107  15.026107   root   <NA>   <NA>   <NA>
#> 125 35.000000 35.000000   0.000000   6.637771   root   <NA>   <NA>   <NA>
#> 126 36.000000 36.000000   0.000000   6.637771   root   <NA>   <NA>   <NA>
#> 127 35.500000 35.000000   6.637771   6.637771   root   <NA>   <NA>   <NA>
#> 128 35.500000 36.000000   6.637771   6.637771   root   <NA>   <NA>   <NA>
#> 129 37.000000 37.000000   0.000000   3.929377   root   <NA>   <NA>   <NA>
#> 130 38.000000 38.000000   0.000000   3.929377   root   <NA>   <NA>   <NA>
#> 131 37.500000 37.000000   3.929377   3.929377   root   <NA>   <NA>   <NA>
#> 132 37.500000 38.000000   3.929377   3.929377   root   <NA>   <NA>   <NA>
#> 133 39.000000 39.000000   0.000000   8.027453   root   <NA>   <NA>   <NA>
#> 134 40.000000 40.000000   0.000000   8.027453   root   <NA>   <NA>   <NA>
#> 135 39.500000 39.000000   8.027453   8.027453   root   <NA>   <NA>   <NA>
#> 136 39.500000 40.000000   8.027453   8.027453   root   <NA>   <NA>   <NA>
#> 137 37.500000 37.500000   3.929377  13.352260   root   <NA>   <NA>   <NA>
#> 138 39.500000 39.500000   8.027453  13.352260   root   <NA>   <NA>   <NA>
#> 139 38.500000 37.500000  13.352260  13.352260   root   <NA>   <NA>   <NA>
#> 140 38.500000 39.500000  13.352260  13.352260   root   <NA>   <NA>   <NA>
#> 141 35.500000 35.500000   6.637771  15.122897   root   <NA>   <NA>   <NA>
#> 142 38.500000 38.500000  13.352260  15.122897   root   <NA>   <NA>   <NA>
#> 143 37.000000 35.500000  15.122897  15.122897   root   <NA>   <NA>   <NA>
#> 144 37.000000 38.500000  15.122897  15.122897   root   <NA>   <NA>   <NA>
#> 145 31.875000 31.875000  15.026107  20.598507   root   <NA>   <NA>   <NA>
#> 146 37.000000 37.000000  15.122897  20.598507   root   <NA>   <NA>   <NA>
#> 147 34.437500 31.875000  20.598507  20.598507   root   <NA>   <NA>   <NA>
#> 148 34.437500 37.000000  20.598507  20.598507   root   <NA>   <NA>   <NA>
#> 149 43.000000 43.000000   0.000000   8.537564   root   <NA>   <NA>   <NA>
#> 150 44.000000 44.000000   0.000000   8.537564   root   <NA>   <NA>   <NA>
#> 151 43.500000 43.000000   8.537564   8.537564   root   <NA>   <NA>   <NA>
#> 152 43.500000 44.000000   8.537564   8.537564   root   <NA>   <NA>   <NA>
#> 153 42.000000 42.000000   0.000000  10.771175   root   <NA>   <NA>   <NA>
#> 154 43.500000 43.500000   8.537564  10.771175   root   <NA>   <NA>   <NA>
#> 155 42.750000 42.000000  10.771175  10.771175   root   <NA>   <NA>   <NA>
#> 156 42.750000 43.500000  10.771175  10.771175   root   <NA>   <NA>   <NA>
#> 157 45.000000 45.000000   0.000000  13.044922   root   <NA>   <NA>   <NA>
#> 158 46.000000 46.000000   0.000000  13.044922   root   <NA>   <NA>   <NA>
#> 159 45.500000 45.000000  13.044922  13.044922   root   <NA>   <NA>   <NA>
#> 160 45.500000 46.000000  13.044922  13.044922   root   <NA>   <NA>   <NA>
#> 161 49.000000 49.000000   0.000000   2.291288   root   <NA>   <NA>   <NA>
#> 162 50.000000 50.000000   0.000000   2.291288   root   <NA>   <NA>   <NA>
#> 163 49.500000 49.000000   2.291288   2.291288   root   <NA>   <NA>   <NA>
#> 164 49.500000 50.000000   2.291288   2.291288   root   <NA>   <NA>   <NA>
#> 165 48.000000 48.000000   0.000000  10.184218   root   <NA>   <NA>   <NA>
#> 166 49.500000 49.500000   2.291288  10.184218   root   <NA>   <NA>   <NA>
#> 167 48.750000 48.000000  10.184218  10.184218   root   <NA>   <NA>   <NA>
#> 168 48.750000 49.500000  10.184218  10.184218   root   <NA>   <NA>   <NA>
#> 169 47.000000 47.000000   0.000000  18.993398   root   <NA>   <NA>   <NA>
#> 170 48.750000 48.750000  10.184218  18.993398   root   <NA>   <NA>   <NA>
#> 171 47.875000 47.000000  18.993398  18.993398   root   <NA>   <NA>   <NA>
#> 172 47.875000 48.750000  18.993398  18.993398   root   <NA>   <NA>   <NA>
#> 173 45.500000 45.500000  13.044922  27.779904   root   <NA>   <NA>   <NA>
#> 174 47.875000 47.875000  18.993398  27.779904   root   <NA>   <NA>   <NA>
#> 175 46.687500 45.500000  27.779904  27.779904   root   <NA>   <NA>   <NA>
#> 176 46.687500 47.875000  27.779904  27.779904   root   <NA>   <NA>   <NA>
#> 177 42.750000 42.750000  10.771175  33.117815   root   <NA>   <NA>   <NA>
#> 178 46.687500 46.687500  27.779904  33.117815   root   <NA>   <NA>   <NA>
#> 179 44.718750 42.750000  33.117815  33.117815   root   <NA>   <NA>   <NA>
#> 180 44.718750 46.687500  33.117815  33.117815   root   <NA>   <NA>   <NA>
#> 181 41.000000 41.000000   0.000000  41.094765   root   <NA>   <NA>   <NA>
#> 182 44.718750 44.718750  33.117815  41.094765   root   <NA>   <NA>   <NA>
#> 183 42.859375 41.000000  41.094765  41.094765   root   <NA>   <NA>   <NA>
#> 184 42.859375 44.718750  41.094765  41.094765   root   <NA>   <NA>   <NA>
#> 185 34.437500 34.437500  20.598507  54.746831   root   <NA>   <NA>   <NA>
#> 186 42.859375 42.859375  41.094765  54.746831   root   <NA>   <NA>   <NA>
#> 187 38.648438 34.437500  54.746831  54.746831   root   <NA>   <NA>   <NA>
#> 188 38.648438 42.859375  54.746831  54.746831   root   <NA>   <NA>   <NA>
#> 189 23.796875 23.796875  44.837933  89.232093   root   <NA>   <NA>   <NA>
#> 190 38.648438 38.648438  54.746831  89.232093   root   <NA>   <NA>   <NA>
#> 191 31.222656 23.796875  89.232093  89.232093   root   <NA>   <NA>   <NA>
#> 192 31.222656 38.648438  89.232093  89.232093   root   <NA>   <NA>   <NA>
#> 193  4.765625  4.765625  77.605024 152.313999   root   <NA>   <NA>   <NA>
#> 194 31.222656 31.222656  89.232093 152.313999   root   <NA>   <NA>   <NA>
#> 195 17.994141  4.765625 152.313999 152.313999   root   <NA>   <NA>   <NA>
#> 196 17.994141 31.222656 152.313999 152.313999   root   <NA>   <NA>   <NA>
#>