Build a data frame

This function converts various objects to a data frame.

Usage

# Default S3 method
fortify_data_frame(data, ..., data_arg = NULL, call = NULL)

Arguments

data

An object to be converted to a data frame.

...

Additional arguments passed to fortify().

data_arg

The argument name for data. Developers can use it to improve messages. Not used by the user.

call

The execution environment where data and other arguments for the method are collected. Developers can use it to improve messages. Not used by the user.

Details

By default, it calls fortify() to build the data frame.

See also