Build a matrix from phylo object

This method allows a phylo object to be directly input into stack_discrete() or circle_discrete(). This makes it possible to add align_phylo() to the stack independently, as align_phylo() requires the layout to have labels.

Usage

# S3 method for class 'phylo'
fortify_matrix(data, ..., data_arg = caller_arg(data), call = NULL)

Arguments

data

A phylo object.

...

These dots are for future extensions and must be empty.

data_arg

The argument name for data. Developers can use it to improve messages. Not used by the user.

call

The execution environment where data and other arguments for the method are collected. Developers can use it to improve messages. Not used by the user.

Value

A one-column matrix where the tip labels are the values, and the row names will also be the tip labels.

See also