Free from alignment

plot_grid will try to align plots by their panels, and align every elements of the plot, following functions romove this restrictions.

free_align: if we want to compose plots without alignment of some panel axes (panel won't be aligned). we can wrap the plot with free_align.

free_border: If we want to compose plots without alignment of the panel borders (but still align the panels themselves), we can wrap the plot with free_border.

free_lab: If we want to compose plots without alignment of the axis title, we can wrap the plot with free_lab.

free_size: Removing the ggplot element sizes when aligning

Usage

free_align(plot, axes = c("t", "l", "b", "r"))

free_border(plot, borders = c("t", "l", "b", "r"))

free_lab(plot, labs = c("t", "l", "b", "r"))

free_size(plot, ...)

Arguments

plot

A ggplot or alignpatches object.

axes

Which axes shouldn't be aligned? Allowed values: "t","l","b",and "r".

borders

Which border shouldn't be aligned? Allowed values: "t","l","b",and "r".

labs

Which axis labs to be free?

...

What ggplot2 elements to remove?

Value

free_align: A modified version of plot with a free_align class.

free_border: A modified version of plot with a free_border class.

free_lab: A modified version of plot with a free_lab class.

free_size: A modified version of plot with a free_size class.