Determine the context of subsequent manipulations

Determine the context of subsequent manipulations

Usage

hmanno(what, size = NULL, guides = NULL, align_axis_title = NULL, plot_data)

Arguments

what

What should get activated? Possible values are follows:

A string of "top", "left", "bottom", or "right".

NULL: means set the active context into the heatmap itself.

missing or NA: don't change the context, use current active context.

If the active context is a string of "top", "left", "bottom", or "right". We can also set the stack (heatmap annotation) size, guides align_axis_title and plot_data.

size

A unit object to set the size of the heatmap annotation. This will only be used if what is a string of "top", "left", "bottom", or "right".

guides

A string specifying how guides should be treated in the layout. "collect" will collect guides below to the given nesting level, removing duplicates. "keep will stop collection at this level and let guides be placed alongside their plot. "auto" will allow guides to be collected if a upper level tries, but place them alongside the plot if not. If you modify default guide "position" with theme(legend.position=...) while also collecting guides you must apply that change to the overall layout.

align_axis_title

A boolean value or a character of the axis position ("t", "l", "b", "r") indicates how to align the axis title. By default, all axis title won't be aligned.

plot_data

A function used to transform the plot data before rendering. By default, it'll inherit from the parent layout. If no parent layout, the default is to not modify the data. Use NULL, if you don't want to modify anything. Used to modify the data after layout preparation has been applied, but before the data is handled of to the ggplot2 for rendering. Use this hook if the you needs change the default data for all geoms.

Value

A active object which can be added into LayoutHeatmap.