Reports whether x is layout object

Reports whether x is layout object

Usage

is_layout(x)

is_quad_layout(x)

is_stack_layout(x)

is_heatmap_layout(x)

is_ggheatmap(x)

Arguments

x

An object to test.

Value

A single boolean value.

Examples

is_layout(ggheatmap(1:10))
#> [1] TRUE

# for quad_layout()
is_quad_layout(quad_alignb(1:10))
#> [1] TRUE
is_quad_layout(quad_alignh(1:10))
#> [1] TRUE
is_quad_layout(quad_alignv(1:10))
#> [1] TRUE
is_quad_layout(quad_free(mtcars))
#> [1] TRUE

# for stack_layout()
is_stack_layout(stack_align(1:10))
#> Warning: The `direction` argument of `stack_align()` must be provided as of ggalign
#> 0.0.5.
#> ℹ no default value of 'direction' argument in the next release
#> ℹ The deprecated feature was likely used in the ggalign package.
#>   Please report the issue at <https://github.com/Yunuuuu/ggalign/issues>.
#> [1] TRUE
is_stack_layout(stack_free(1:10))
#> Warning: The `direction` argument of `stack_free()` must be provided as of ggalign
#> 0.0.5.
#> ℹ no default value of 'direction' argument in the next release
#> ℹ The deprecated feature was likely used in the ggalign package.
#>   Please report the issue at <https://github.com/Yunuuuu/ggalign/issues>.
#> [1] TRUE

# for heatmap_layout()
is_heatmap_layout(ggheatmap(1:10))
#> [1] TRUE
is_ggheatmap(ggheatmap(1:10))
#> [1] TRUE