Put plots horizontally or vertically

ggstack is an alias of layout_stack.

Usage

layout_stack(
  data,
  direction = NULL,
  rel_sizes = NULL,
  guides = NULL,
  align_axis_title = NULL,
  plot_data = waiver()
)

ggstack(
  data,
  direction = NULL,
  rel_sizes = NULL,
  guides = NULL,
  align_axis_title = NULL,
  plot_data = waiver()
)

Arguments

data

A numeric or character vector, a data frame, or a matrix.

direction

A string of "horizontal" or "vertical", indicates the direction of the stack layout.

rel_sizes

A numeric or unit object of length 3 indicates the relative widths (direction = "horizontal") / heights (direction = "vertical").

guides

A string specifying how guides should be treated in the layout. "collect" will collect guides below to the given nesting level, removing duplicates. "keep will stop collection at this level and let guides be placed alongside their plot. "auto" will allow guides to be collected if a upper level tries, but place them alongside the plot if not. If you modify default guide "position" with theme(legend.position=...) while also collecting guides you must apply that change to the overall layout.

align_axis_title

A boolean value or a character of the axis position ("t", "l", "b", "r") indicates how to align the axis title. By default, all axis title won't be aligned.

plot_data

A function used to transform the plot data before rendering. By default, it'll inherit from the parent layout. If no parent layout, the default is to not modify the data. Use NULL, if you don't want to modify anything. Used to modify the data after layout preparation has been applied, but before the data is handled of to the ggplot2 for rendering. Use this hook if the you needs change the default data for all geoms.

Value

A LayoutStack object.