Modify - Operator Context in quad_layout()

Wrapping objects with with_quad allows you to change the context of the subtraction - operator in quad_layout().

Usage

with_quad(x, position = waiver(), main = NULL)

Arguments

x

Objects to add to the layout using the - operator.

position

A string containing one or more of "t", "l", "b", and "r" indicating which annotation stack should be used as the context. If NULL, the context will default to the quad_layout() itself. By default, waiver() is used, which triggers the following behavior: If the current active context in quad_layout() is top or bottom, the operator will also act on the corresponding bottom or top annotation. If the context is left or right, the operator will also act on the right or left annotation, respectively. If there is no active annotation stack, it defaults to NULL.

main

A single boolean value indicates whether to apply x for the main plot.

Value

The input object with an additional attribute that specifies the selected context.