Create Align object

Align object will act with the layout object, reorder or split the observations, some of them can also add plot components into the layout object.

Usage

align(
  align_class,
  params,
  size = NULL,
  data = NULL,
  plot_data = waiver(),
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  check.param = TRUE,
  call = caller_call()
)

Arguments

size

Plot size, can be a unit object.

data

A matrix, a data frame, or even a simple vector that will be converted into a one-column matrix. If the data argument is set to NULL, the align_* will use the layout data. Additionally, the data argument can also accept a function (purrr-like lambda is also okay), which will be applied with the layout data,

It is important to note that all align_* functions consider the rows as the observations. It means the NROW(data) must return the same number with the parallel layout axis.

layout_heatmap: for column annotation, the layout data will be transposed before using (If data is a function, it will be applied with the transposed matrix). This is necessary because column annotation uses heatmap columns as observations, but we need rows.

layout_stack: the layout data will be used as it is since we place all plots along a single axis.

plot_data

A function used to transform the plot data before rendering. By default, it'll inherit from the parent layout. If no parent layout, the default is to not modify the data. Use NULL, if you don't want to modify anything. Used to modify the data after layout preparation has been applied, but before the data is handled of to the ggplot2 for rendering. Use this hook if the you needs change the default data for all geoms.

limits

A boolean value indicates whether to set the layout limtis for the plot.

facet

A boolean value indicates whether to set the layout facet for the plot. If this is FALSE, limits will always be FALSE too.

set_context

A single boolean value indicates whether to set the active context to current Align object. If TRUE, all subsequent ggplot elements will be added into this Align object.

order

An single integer for the layout order.

name

A string of the object name.

check.param

A single boolean value indicates whether to check the supplied parameters and warn.

call

The call used to construct the Align for reporting messages.

Value

A new Align object.

Align

Each of the Align* objects is just a ggproto() object, descended from the top-level Align, and each implements various methods and fields.

To create a new type of Align* object, you typically will want to override one or more of the following:

setup_params: Prepare parameter or check parameters used by this annotation.

setup_data: Prepare data used by this annotation.

compute: A method used to compute statistics.

layout: A method used to group heamap rows/columns into panel or reorder heamtap rows/columns.

draw: A method used to draw the plot. Must return a ggplot object.