Build a matrix from a maftools object

Convert MAF object to a matrix:

fortify_matrix.MAF: Extract genomic alterations for genes.

fortify_matrix.MAF_pathways: Extract genomic alterations for pathways. tune.MAF() helps convert MAF object to a MAF_pathways object.

Usage

# S3 method for class 'MAF'
fortify_matrix(
  data,
  ...,
  genes = NULL,
  n_top = NULL,
  remove_empty_genes = TRUE,
  remove_empty_samples = TRUE,
  collapse_vars = TRUE,
  use_syn = TRUE,
  missing_genes = "error",
  data_arg = caller_arg(data),
  call = NULL
)

# S3 method for class 'MAF_pathways'
fortify_matrix(
  data,
  ...,
  pathdb = "smgbp",
  remove_empty_pathways = TRUE,
  remove_empty_samples = TRUE,
  data_arg = caller_arg(data),
  call = NULL
)

Arguments

data

A MAF object.

...

These dots are for future extensions and must be empty.

genes

An atomic character defines the genes to draw.

n_top

A single number indicates how many top genes to be drawn.

remove_empty_genes

A single boolean value indicats whether to drop genes without any genomic alterations.

remove_empty_samples

A single boolean value indicats whether to drop samples without any genomic alterations.

collapse_vars

A single boolean value indicating whether to collapse multiple alterations in the same sample and gene into a single value "Multi_Hit". Alternatively, you can provide a single string indicates the collapsed values.

use_syn

A single boolean value indicates whether to include synonymous variants when Classifies SNPs into transitions and transversions.

missing_genes

A string, either "error" or "remove", specifying the action for handling missing genes.

data_arg

The argument name for data. Developers can use it to improve messages. Not used by the user.

call

The execution environment where data and other arguments for the method are collected, e.g., caller_env(). Developers can use it to improve messages. Not used by the user.

pathdb

A string of "smgbp" or "sigpw", or a named list of genes to define the pathways.

remove_empty_pathways

A single boolean value indicats whether to drop pathways without any genomic alterations.

ggalign attributes

For fortify_matrix.MAF:

gene_summary: gene summary informations. See maftools::getGeneSummary() for details.

sample_summary: sample summary informations. See maftools::getSampleSummary() for details.

sample_anno: sample clinical informations. See maftools::getClinicalData() for details.

n_genes: Total of genes.

n_samples: Total of samples.

titv: A list of data.frames with Transitions and Transversions summary. See maftools::titv() for details.

For fortify_matrix.MAF_pathways:

gene_list: the pathway contents.

pathway_summary: pathway summary informations. See maftools::pathways() for details.

sample_summary: sample summary informations. See maftools::getSampleSummary() for details.

sample_anno: sample clinical informations. See maftools::getClinicalData() for details.

See also