Determine the active context of heatmap layout

Determine the active context of heatmap layout

Usage

hmanno(
  position = NULL,
  size = NULL,
  width = NULL,
  height = NULL,
  guides = NA,
  free_labs = NA,
  free_sizes = NA,
  plot_data = NA,
  theme = NULL,
  what = waiver()
)

Arguments

position

Which heatmap annotation should get activated? Possible values are follows:

A string of "top", "left", "bottom", or "right".

NULL: means set the active context into the heatmap itself.

size

A unit object to set the total size of the heatmap annotation. This will only be used if what is a string of "top", "left", "bottom", or "right".

If position is "top" or "bottom", size set the total height of the annotation.

If position is "left" or "right", size set the total width of the annotation.

width, height

Heatmap body width/height, can be a unit object. Only used when position is NULL.

guides

A single boolean value indicates whether to collect the guides, or you can specify the string of the guide position to collect. Allowed strings are: "t","l","b",and "r".

free_labs

A boolean value or a character of the axis position ("t", "l", "b", "r") indicates which axis title should be free from alignment. By default, all axis title won't be aligned.

free_sizes

A character specifies the ggplot elements which won't count space sizes when alignment.

plot_data

A function used to transform the plot data before rendering. By default, it'll inherit from the parent layout. If no parent layout, the default is to not modify the data. Use NULL, if you don't want to modify anything.

Used to modify the data after layout has been created, but before the data is handled of to the ggplot2 for rendering. Use this hook if the you needs change the default data for all geoms.

theme

A theme() object to rendering the guides, margins and background. Only used when position is NULL.

what

What should get activated for the stack layout? Possible values are follows:

A single number or string of the plot elements in the stack layout. Usually you are waive to use this, since the adding procedure can be easily changed.

NULL: Remove any active context, this is useful when the active context is a LayoutHeatmap, where any Align objects will be added into the heatmap. By removing the active context, we can add Align object into the StackLayout.

Value

A active object which can be added into HeatmapLayout.