Modify components of the layout theme

Used to render the guides, title, subtitle, caption, margins, panel.border background.

Usage

layout_theme(...)

Arguments

...

Additional arguments passed to theme().

Value

A layout_theme object to be added to heatmap_layout() or stack_layout() object.

Theme inheritance

Theme elements inherit properties from other theme elements hierarchically. For example, axis.title.x.bottom inherits from axis.title.x which inherits from axis.title, which in turn inherits from text. All text elements inherit directly or indirectly from text; all lines inherit from line, and all rectangular objects inherit from rect. This means that you can modify the appearance of multiple elements by setting a single high-level component.

Learn more about setting these aesthetics in vignette("ggplot2-specs").

Examples

layout_theme(plot.background = element_rect(fill = "green"))
#> List of 1
#>  $ plot.background:List of 5
#>   ..$ fill         : chr "green"
#>   ..$ colour       : NULL
#>   ..$ linewidth    : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  - attr(*, "class")= chr [1:3] "layout_theme" "theme" "gg"
#>  - attr(*, "complete")= logi FALSE
#>  - attr(*, "validate")= logi TRUE